refactor: move `publish` out of realism_authoring into room_qc
`publish` ran last in the authoring flow, so it lived under `realism_authoring`.
But running last is not the same as belonging: by the time it runs the room is
finished being built. What it actually does is gate the room — collision
correction, then the quality report — and produce the artifacts a human judges
it by (the compiled glb, the baked materials, the real-vs-render comparison).
That is quality control and presentation, not authoring.

Two symptoms of it being in the wrong place:

  * it already called into `room_qc` for both of its first two steps, reaching
    across packages to do the thing it was mostly made of;
  * `cli.py`'s `view` command imports `publish.viewable_room` purely to find
    out where a room's glb is — viewing has nothing to do with publishing, so a
    stage was being used as a library because that is where the helper lived.

Pure move: the stage still runs after `author` because `stages.py` declares
that dependency explicitly rather than inferring it from the module's location.
No behaviour change.
