+++ tests/test_walk_viewer.py
@@ -15,8 +15,8 @@ from urllib.request import urlopen
 import pytest
 
 from litereality_agent.pipeline.context import RunContext
-from litereality_agent.pipeline.realism_authoring import publish
-from litereality_agent.pipeline.realism_authoring.publish import WEB_GLB_NAME
+from litereality_agent.pipeline.room_qc import publish
+from litereality_agent.pipeline.room_qc.publish import WEB_GLB_NAME
 from litereality_agent.room_ops import walk
